+++ CHANGELOG.md
@@ -1,13 +1,16 @@
 # Changelog
 
 ## [Unreleased]
+* Changes
+  * Removed the `get_user_profile` method and the `email_address` and `profile` attributes from the `Auth` class ([#421](https://github.com/nsidc/earthaccess/issues/421))
 * Bug fixes:
     * fixed 483 by extracting a common CMR query method for collections and granules using SearchAfter header
     * Added VCR support for verifying the API call to CMR and the parsing of returned results without relying on CMR availability post development
 
 * Enhancements:
   * Corrected and enhanced static type hints for functions and methods that make
-    CMR queries or handle CMR query results (#508)
+    CMR queries or handle CMR query results ([#508](https://github.com/nsidc/earthaccess/issues/508))
+  * Enable queries to Earthdata User Acceptance Testing (UAT) system for authenticated accounts ([#421](https://github.com/nsidc/earthaccess/issues/421))
 
 ## [v0.9.0] 2024-02-28